requirements.txt
Django==2.2.6
django-crispy-forms==1.7.2
django-tempus-dominus==5.1.2.8
Pillow==6.2.0
psycopg2==2.8.3
pytz==2019.3
sqlparse==0.3.0
